Add some implementation of RelocData and MCLinker.
Sorry I cannot write a unittest, since MCLDInfo and TargetLDBackend
is not done yet, and MCLinker needs both of these.

This is a placeholder commit that tell others I have started doing
the relocation part. The function is not done yet.

BTW, the LLVM patch has a bug.
We use `delete' in MCAssembler.h, so we should use `include' instead
of forward declaration.

Be care of compilation warning!!