Change the roles of SectLinker, MCLDInfo, and LLVMTargetMachine.
MCLDInfo comprises of
 - attribute factory
 - input factory
 - context factory
 - the default output
 - script options
 - general options
 - input tree.

The construction of MCLDInfo is target-independent.
Although MCLDInfo contains some target-dependent information, such as the
default attribute and the default script, the construction of these data
should be target-independent. Therefore, we move the construction of MCLDInfo
to LLVMTargetMachine.

SectLinker now is in charge of the construction of target-indepedent data.
Such as the default input file and the default output file. The target-
dependent data are contructed by the inheritor of SectLinker. Take
ARMELFSectLinker for example, target depedent data, like default link script
and the defaut attribute are configured in the constructor of ARMELFSectLinker.