Fix a bug - all MCLDInput get the same attribute.
For easilier readable, I rename
  MCLDAttribute -> AttributeProxy
  AttributeBase -> Attribute

Since many mcld::Input shares common attribute, but AttributeFactory
can not assign the very identical pointer of attribute to different
mcld::Input when the pointed attribute will be modified in the future.

So, we let AttributeProxy hold the pointer of the attriubte. When we
change the attributes by proxy, mcld::Input will be affected and stick
to the same Attribute.