Add MallocAllocator
LLVM has provided a llvm::MallocAllocator, but it is a 'plain' allocator.
That is, llvm::MallocAllocator returns a raw byte string and users need to
cast the raw byte string on a new type by itself. Additionally,
llvm::MallocAllocator does not follow the API defined in C++ standard.
So, I decide to give them a new MallocaAllocator.