Refactoring HashTable
1. make HashTable is no longer limited to llvm::StringRef key type.

HashTable get the type of key and value from HashEntry.
HashEntry::key_type and HashEntry::value_type. HashTable does not assume the key type is based on string. Moreover, the hash function is also not limited to the string hash.

2. move the construction of HashEntry outside from the HashTable.

Clients must pass the factory of HashEntry as a template argument into the HashTable. That is,

  HashTable<..., EntryFactory>

The memory can control by the factory. HashTable no longer takes care of it.