1. Change the HashTable from quadratic probing to linear probing. 2. Fix bugs in ChainIteratorBase
At first, we follow llvm, use quadratic probing. However, in llvm, the quadratic function is not real quadratic, it's a triangle function - K(K+1)/2. Some number of buckets, even when the load factor is under the threshold, can not be probed by given triangle function. This will cause a infinite loop in HashBase::lookUpBucketFor. In order to avoid the problem, I change the hash table to linear probing.

ChainIteratorBase::advance behave wrong. It returned when meeting a empty bucket. It should ignore empty bucket.