LDSection should be a kind of MCSection.
LDSection has name, type, flags and offset of a section.

MCLinker reuses llvm::MCSectionData. llvm::MCSectionData only contains
partial information of a section, such as real data size and alignment
constraint. We complement the rest information by put them in LDSection.

LDSection is designed for both ELF and MachO. We do not consider about
WinPE when we designed it. So, WinPE may not use it.