Add ResolveInfo::hasPLT() and setHasPLT()
Which comes first? Applying relocations or creating GOT/PLT? This is
like checkens and eggs. Without applying relocation, we don't know
the number of entries of GOT and PLT. But without GOT and PLT's offset,
we can not fix symbol values and applying relocation.

In order to solve this problem, MCLinker separate the creating GOT/PLT process
into two different phases. The first phase is determining if the relocation
entry will produce a GOT and a PLT entry, then *create empty* GOT and
PLT whose sizes are precisely the same with the final sizes. The second phase
is applying the relocation entries, and simultaneously *fill* the GOT and PLT.

ResolveInfo::hasPLT() and setHasPLT() are functions for the relocation entry which
refer to the symbol. If a relocation entry needs to create an empty entry, than
the symbol is setHasPLT(true). And when applying the relocation, MCLinker looks
up the hasPLT(). If it's true, MCLinker will fill the value of the PLT entry and
GOT entry.

By this two-phase solution, we have no performance lost, and still keep
applying relocation and file layout simply.

Google gold scans relocation entries in file many times to deal with this problem.
But we know the scale of the number of relocation entries are huge, and it is
performance critical for on-device linking.