Add API - TargetLDBackend::scanRelocation(Relocation&)
Determine the entries such as GOT and PLT are needed or
not for the input relocation. If needed, create the
empty entries.

After symbol resolution has been done, MCLinker traveses
relocations to generate empty entries. These empty entries
are genrtated for Layout to determine the ouput offset.

For ARM and X86, GOT, PLT and dynamic relocation entries
are created if needed.

For Mips, GOT, dynamic relocation, GP, or other target
dependent entrires are created here if needed.