Fix bugs - There are some relocation entries point to the undefined symbol, and undefined symbol has no fragment reference. These undefined fragment references cause segmentation fault when applying relocations.
The bug source is: in ELF, the relocations' fragment references
come from relocation section's sh_info, not from the dynamic
symbols' st_shndx. So, ELFReader is responsible for set the
LDSection::setInfoLink() when reading relocation sections.

The same problem is easy in MachO, because MachO's relocation table
is within the section. There is no extra effort to get the relocations'
fragment references.