Change the defintion of MCLinker::defineSymbol.
Because some symbols need to set as a local symbol forcefully, so we
should separate MCLinker::defineSymbol() into two functions:

  MCLinker::defineSymbol<AsNeed>() and
  MCLinker::defineSymbol<Force>()

There two functions are both inserting symbol into the symbol table and
resolve the symbol immediately.

MCLinker::defineSymbol<AsNeed> is just like Google gold linker's only_if_ref.
If there is existing an old symbol, the symbol is not defined until the
old symbol is a undefined reference.

MCLinker::defineSymbol<Force>() defines the symbol and resolves it, no matter
if the existing symbol is a undefined reference.