Fix a bug - do not report error when applying relocation with branch type if its target symbol's reserved() is not 0 or PLT
Because there might be a symbol with GOT and there is another
relocation with branch related relocation type
(CALL, PLT32, ...) point to this symbol, and we no need to
generate PLT entry for it. In this situation, the symbol's
reserved() won't be 0 or PLT, but it's still correct and we
should perform static relocation on it.