Avoid summing up sizes of .ARM.attributes sections of input files for output.
MCLinker uses the .ARM.attributes section of the first input file for output.
The size of .ARM.attributes in output is same as the one in first input file.
there is no need to sum up size for ARM.attributes section.

MCLinker will emit .ARM.attributes section in correct way in the future.