Modify API - TargetLDBackend::scanRelocation
Add a parameter pInputSym which is the input LDSymbol of
relocation target symbol.

We need to update value in relocation place (A value) if it's a
ABS type relocation and point to a section symbol. This value
should be the offset to the output section. We needs the
input section symbol to get the correct output offset if several
sections are merged to single one.