Add allocateCommonSymbols to TargetLDBackend.
Although different formats allocate common symbols in different section, but
all formats have steps to allocate common symbols. So, we add a new API
TargetLDBackend::allocateCommonSymbols() and implement it in different targets.

Some targets, such as Mips and X86_64, have different need to allocate
common symbols. Their LDSection, the definition of common symbols, etc. are
very different. So, we leave the implementation of allocateCommonSymbols() to
the lowerest target backends.

In ARM and X86, a basic allocation is implemented. But Mips needs small symbol
support.