Separate the implementation of LEB128 from header.
No functionality or API changed.

Fully specialized template functions are just like the normal
functions. It doesn't depend on the template parameters any more.
Therefore, you should declare them as inline functions in header
or define them in a .cpp in order to not violating the ODR. And
we choose the later.