Add a HashMap into GNULDBackend to map symbols to their ouput symtab index
1. Add a HashMap m_pSymIndexMap, maintain the map when emitDynNamePools
   and emitRegNamePools
2. Add an API GNULDBackend::getSymbolIdx to get symbol's index. This
   API is used when emitRelocation. This API takes the place of
   original LDContext::getSymbolIdx.

Now only the input symbols will push into LDContex's symTab, the
output symbols and the index maintained in m_pSymIndexMap of GNULDBackend.

When emitRelocation, we get the symbol's index in ouput symbol table
by asking the HashTable, this can reduce time on searching from the
vector and avoid string comparison.