Add API Layout::appendFragment().
To handle alignment in section merging correctly, we need to preserve input
alignment constraint. For mergeable sections (e.g., .text, .data, and .bss, ...)
from input objects, we should use this new API to append the fragment to the
associated MCSectionData.

As for the backend part, there is no need to change the used API.