Add COPY relocation support to ARM backend
When building non-pic executable object, a copy relocation
may need for symbol defined in shared object. Three functions
are added to support copy relocation.

1. ARMLDBackend::needCopyReloc - check if the symbol needs
   a copy relocation or not.

2. ARMLDBackend::defineSymbolforCopyReloc - if the symbol
   needs a copy relocation, defined a copy symbol in the
   BSS section.

3. ARMLDBackend::addCopyReloc - add a copy relcation into
   .rel.dyn which aginst the copy symbol in the bss section.

NOTICE: Append a fragment into BSS for the copy symbol here
will cause the error report in Layout.cpp, line 139, which
is not suitable error check for this case.