Add `External' mcld::Input type.
MCLinker is in the direction to be a modular linker.

MCLinker is stong on object format. The object intermediate representation (OIR)
in MCLinker is designed for various targets and formats. All formats and targets
can use the identical IR to perform linking. Linking algorithm is format and
target independent.

Let our customers define their own input file and format is important. So, we
add a new `External' type of mcld::Input, and our customers can define their own
input file.