Fix a bug - Issue 78 - d_D, U should not override the symbol.
plasma.o:
    55: 00000000     0 OBJECT  GLOBAL DEFAULT  UND _gp_disp

plasma.so:
     1: 0000aba4     0 SECTION GLOBAL DEFAULT  ABS _gp_disp
    30: 0000aba4     0 SECTION GLOBAL DEFAULT  ABS _gp_disp

o (U), so (d_D): should be
use the dynamic definition if we have a reference. rise isDynamic, then change
back the symbol as undefined as emission.

so (d_D), o (U): should be
remember the binding of the undefined. Do not override the symbol. rise isDynamic.