By Hemant Kulkarni's report, add missing headers.
--
Are some of the patches missing from source:
mclinker/optimized/../include/mcld/Object/ObjectLinker.h:122: error: 'NULL' was not declared in this scope

This is not the head but the revision mentioned in the getting started instruction page. Is the tip of mclinker repo buildable? Or what is the last known good source to build ?