Introduce to CodePosition concept.
CodePosition is a novel concept. BFD and Google gold linker do not have this.

In the MCLinker implementation, CodePosition indicates the ability of the
generated output to be loaded at different addresses. There are three
categories in CodePosition: Independent, DynamicDependent, and StaticDependent.

If the output can be loaded at different addresses, we say the output is
position independent. Shared libraries and position-independent executable
programs (PIE) are in this category.

If a executable program can not be loaded at arbitrary addresses, but it
can call outside functions, we say the program is dynamic dependent on
the address to be loaded. In general, executable programs are dynamic
dependent.

If a executable program can not be loaded at different addresses, and
only call inner functions, then we say the program is static dependent on
its loaded address.

LinkerConfig::codePosition() returns the CodePosition of the output.

This patch removes FragmentLinker::isOutputPIC() and isStaticLink(), and
replace the callside by LinkerConfig::codePosition().