Path Conflict
On a Mac Mini with an Ubuntu 12.04 operating system, there was a path conflict when the make install command
was executed.  If using a relative path, the file path in both target and  source directory conflicted.
In order to resolve this problem, it was necessary to use an absolute path instead of a relative path.