[Attribute] Skeleton data structures.
This and the subsequent commits aim to support attribute sections in
ELF. More specifically, we are going to handle .ARM.attribute
correctly.

The file format of attribute section were developed by ARM and the
document can be found in:

"Addenda to, and Errata in, the ABI for the ARM Architecture"
http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.ihi0045d/index.html
"Addenda to, and Errata in, the ABI for the ARM Architecture"

which is referred as "ARM [ABI-addenda]" in the code.

ELFAttribute manages attribute data in the output section. Target
backends use registerAttributeData() to place their data under
ELFAttribute's management.

It keeps merging attributes from the input one by one to the
corresponding ELFAttributeData. ELFAttributeData stores the real
data, providing APIs to query the attribute value and implements the
logics of how to merge the attribute as per specification.