Refine EhFrame::getDataStartOffset().
EhFrame entry has 32-bit and 64-bit version, but it is not according to
the object file 32 or 64. Actually, even 64-bit object file still uses
32-bit eh_frame entries.

We refine getDataStartOffset() to a class template function, so that we don't
need to set it per input and don't need a static member to keep it.