Refinement: Remove those invalid relocations.
If we don't want some relocations to be processing, we remove relocations from
the fragment list instead of setting them ignored. This is useful when someone
is dropped in the middle stage but some relocations still point to it.

Note that we don't need to delete the relocations since they are created
by GCFactory.