Avoid possible use-after-delete.
The evaluation ordering of the expression `erase(it++)` is unspecified.
Thus, it is possible that the iterator is first deleted and then
increased, which will result in an use-after-delete.

The correct way is to make sure it++ is sequenced before the `erase()`
member function call so that the iterator won't be invalidate before the
incremental.